Revert "parseopts: normalize options into an array"
This was really only half a fix for FS#28445, as it still doesn't
correctly handle the case of filenames with spaces. In the short term,
there is no obvious fix for this. In the long term, I believe the
correct decision is to rewrite the options parser to be more in line
with GNU getopt_long.

This reverts commits:
  ca4142714137b16feabac09c4cda86b0a75036f8.
  969dcddbdf9d5dbd91aa414cdd193f3fb26b644b.